requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # For data models validation and settings

# Media processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3
opencv-python>=4.8.0
scenedetect>=0.6.2

# Audio processing
pyannote.audio>=3.1.0
torch>=2.0.0
librosa>=0.10.0
faster-whisper>=1.0.0
openai-whisper>=20231117

# LLM APIs
openai>=1.0.0  # or anthropic>=0.7.0

# Utilities
numpy>=1.24.0
Pillow>=10.0.0
click>=8.1.0  # For CLI
tqdm>=4.66.0  # For progress bars
faiss-cpu>=1.8.0  # Vector similarity search backend for Stage 2

# AWS SDK
boto3>=1.34.0  # AWS SDK for S3, SQS, Lambda operations

# Vector Database
pinecone>=5.0.0  # Pinecone vector database client (v5+ uses new API)

# API Framework (Stage 3.3)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0  # ASGI server (for local dev)
python-multipart>=0.0.6  # For file uploads
mangum>=0.17.0  # AWS Lambda adapter for ASGI applications

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For async HTTP testing with FastAPI